API around config now works in a more intuitive way (#79)
Ended up changing how the extract settings worked a bit. Now there's a custom validator to ensure that the value doesn't come in as `null`. This mostly forces the API to act more like just passing human-ish values to the config.yaml.

I ran into some fun issues due to https://github.com/samuelcolvin/pydantic/issues/990, so went with it defaulting to a custom class that I end up manually removing from the model before dumping back out as JSON or YAML.

Also, now I have a lot of the fields expressed in code. Going to try creating dynamic forms (and maybe build the settings in the docs later). 

Still a lot of work to do in the frontend. This PR was supposed to be for that, but a lot more backend issues kept cropping up!