UI: Removed broadcaster and switched to single process (#95)
broadcaster caused some issues with local development using the memory module. If we ever switch to an external manager, I'll probably try it again. I've also gone down to one UI process, this was mostly to avoid the need of having redis/postgres also running just so the websockets would work reliably if people were on different server instances. 

Also, since the UI will likely only be used by a single person at a time, and some of the methods are async (logging/websockets, and serving static content), we will likely be fine. And more of the methods could be also be made async (rhasspy api client, and some of the file access).